Initial Commit
I think, that writing small cool TUI projects is the best way for me to
learn Rust, actually practicing it.

I plan to make a terminal-like app, where the commands user types and
the output that is recieved are independed from each other. I want to
see many commands I wrote at the same time (and using the up-arrow key
the whole time doesn't cut it for me) and only see the output of the
latest one.

In this commit I just made a simple layout with `ratatui`, and you can
exit the code pressing the key q. I basically followed the Counter App
tutorial, slightly modifying the the layout. The link to the tutotial is
here: https://ratatui.rs/tutorials/counter-app/