Updated the Layout
I made sure that output and input take the most amoount of area on the
scree, and errors are below. I also in future may add a key-sequence
that would collapse the error area altogether, and another one, that
would bring it back.

I do not think I did anything good. I think you are not supposed to
create two chunks arrays, there is a way to style the layout in a
simpler way, without creating and invisible block. But I do not know
how. Also the only thing that I did so far is some frontend stuff, I
didn't write a single line of logic. I need to continue, actually making
this good.