Added Write Text in Input Section Functionality
A user can now write text in the Input Section of the Spliminal. User
can also delete the text. Also you can only exit from the app if you
haven't selected an active block, but I may change it so a user can exit
from any Block that is not input.