Added Functionality to Execute Commands
I just used the `std::process::Command;` from rust, and it worked! I am
honestly amazed, how simple and engaging the whole process was. I should
read the docs about the topic even more, because now I treat
Command::new() as a magic black box. I do not process the error now,
because I am not sure what is it. Is it like the standart error stream?
Like the one that has something to do with 1> and 2> thingies?

It is crazy, how yesterday evening I thought of this idea, and now I
have a working prototype, crazy. I am very emotional right now.