Added Multi-Line functionality
So now a user can type many commands, and see what is the current number
of command. Although, in each of the sections I just map over the
Strings and display an index. Of course this is not informative, because
a command from the inptut can go to either error block (another thing
that I have added) or an output block. I need to store it better, and
think of the logic more.

Also, I read more about how I execute commands. Now I just do `sh
<command>`, which is the Bourne shell I spawn each time I want to
execute the command. Of course this is no good, because I cannot like
`cd` to a different directory in one command and then see what is inside
it from another command. Although I can achieve the same behaviour by
doing: `cd ./some-dir && ls`. I need to read more about this, but I am
honestly confused.

Another point is that I have practiced with `.iter()` and wrote a
closure, two "cool" Rust features.